requirements.txt
azure_storage
openai
python-dotenv
tenacity
tiktoken